+++ book_learningCpp/ch02/main5.cpp
@@ -1,0 +1,8 @@
+#include <iostream>
+
+int main()
+{
+    unsigned int i{};
+    i = -16;
+    std::cout << i << std::endl;
+}
