--- ch02/includes/various.h
@@ -1,8 +1,0 @@
-#ifndef VARIOUS_H
-#define VARIOUS_H
-
-void action1();
-bool accept3();
-void examineReference();
-
-#endif // VARIOUS_H
